리멤버 미 기능 보수

+++ client/src/views/community/Article.vue
@@ -1,58 +1,75 @@
 <template>
-  <div class="article">
+  <div class="article my-5">
     <div class="container">
-    <div class="card">
-        <div class="card-body">
-            <h2 class="card-title" align="start">제목: {{ this.article.title }}</h2>      
+
+      <div>
+        <h2 class="text-start" align="start"><strong>{{ this.article.title }}</strong></h2>      
+      </div>
+      <!-- 작성자, 작성일, 최종 수정 시간.. -->
+      <div class="row justify-content-start  align-items-center mt-5">
+        <div class="col-1 text-start username">
+          {{ this.article.user }}
         </div>
-        <div class="card-body">
-            <h2 align="start">내용: {{ this.article.content }}</h2>    
-        </div>     
-        <div class="card-body">
-            <h5 align="left">생성시간: {{ this.article.created_at }}</h5>  
-            <h5 align="left">수정시간: {{ this.article.updated_at }}</h5>
-        </div>   
-    </div>
-    <div class="row mt-3">
-      <div class="col-auto">
+        <div class="col-2 text-start" style="color: lightslategray;">
+          last update: {{ this.article.updated_at.substring(0, 10) }}
+        </div>
+      </div>
+        
+      <div class="mt-5 text-start">
+        <span style="font-size: 40px;">{{ this.article.content.substring(0,1) }}</span>
+        {{ this.article.content.substring(1,) }}
+      </div>     
+      <div class="mt-3" align="right" style="color: lightslategray;">
+        craeted at: {{ this.article.created_at.substring(0, 10) }}
+      </div>   
+
+      <div class="row mt-3">
+        <div class="col-auto">
           <button type="button"
             v-if="article.user === user_id"
-            @click="toUpdatePage"  
+            @click="$router.push({name: 'Update'})"  
             class="btn btn-outline-secondary">
             게시글 수정
           </button>
-      </div>
-      <div class="col-auto mb-5">
+        </div>
+        <div class="col-auto">
           <form>                
-              <button type="button"              
-                v-if="article.user === user_id"
-                @click="onDelete"  
-                class="btn btn-outline-secondary">
-                게시글 삭제
-              </button>
+            <button type="button"              
+              v-if="article.user === user_id"
+              @click="onDelete()"  
+              class="btn btn-outline-secondary">
+              게시글 삭제
+            </button>
           </form>
+        </div>
+        <div class="col-auto">
+          <button class="btn btn-outline-secondary" @click="$router.push({name: 'Community'})">목록으로</button>
+        </div>
       </div>
-       </div>
     </div>
+
+    <!-- 댓글 부분 --> 
     <div class="mb-4">
-        <h3 >댓글 작성</h3>
-        <input name="comment" id="comment" v-model="comment_content"><button @click="commentSubmit" class="btn btn-outline-secondary">작성</button>        
+      <h3 >댓글 작성</h3>
+      <input name="comment" id="comment" v-model="comment_content"><button @click="commentSubmit()" class="btn btn-outline-secondary">작성</button>        
     </div>
+
     <div class="container">
-    <div class="card">
-        <div class="card-body">
-           <div v-for="comment in comments" :key="comment.id">
-         <h6 align="start"> {{ comment.user }} : {{ comment.content }}
-          생성시간: {{ comment.created_at }} 수정시간: {{ comment.updated_at }}
-          <button 
-            v-if="comment.user === user_id"
-            @click="onDelete_Comment(comment)" 
-            class="btn btn-outline-secondary">
-            X
-          </button></h6>
-      </div>   
-        </div> 
-    </div>
+      <div class="card">
+          <div class="card-body">
+            <div v-for="comment in comments" :key="comment.id">
+              <h6 align="start"> {{ comment.user }} : {{ comment.content }}
+                생성시간: {{ comment.created_at }} 수정시간: {{ comment.updated_at }}
+                <button 
+                  v-if="comment.user === user_id"
+                  @click="onDelete_Comment(comment)" 
+                  class="btn btn-outline-secondary">
+                  X
+                </button>
+              </h6>
+            </div>   
+          </div> 
+      </div>
     </div>
   </div>
 </template>
@@ -103,6 +120,11 @@ export default {
       })
     },
     onDelete: function () {
+      const conf = confirm('게시글을 삭제하시겠습니까?')
+      
+      if (!conf) {
+        return
+      }
       axios({
         url: `${this.$store.state.SERVER_URL}/community/${this.$store.state.article.id}/article_update_delete/`,
         method: 'DELETE',
@@ -118,10 +140,12 @@ export default {
         console.log(err)
       })
     },
-    toUpdatePage: function () {
-      this.$router.push({name: 'Update'})
-    },
     commentSubmit: function () {
+      if (!localStorage.getItem('jwt')){
+        alert('로그인이 필요합니다.')
+        this.$router.push({name: 'Login'})
+      }
+
       const comment = {
         content : this.comment_content,
       }
@@ -163,5 +187,8 @@ export default {
 </script>
 
 <style>
-
+.username{
+  color: dodgerblue;
+  font-size: 20px;
+}
 </style>
